backend/requirements.txt
FastAPI==0.104.1
uvicorn==0.24.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
google-generativeai==0.3.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
